+++ R/utility.R
@@ -387,3 +387,18 @@ get_first_line <- function(x) {
   first_lines <- vapply(split, function(x) x[1], character(1))
   return(first_lines)
 }
+
+# Check for `site: workflowr::wflow_site` in index.Rmd
+check_site_generator <- function(index) {
+  if (!fs::file_exists(index))
+    stop(glue::glue("Unable to find index.Rmd. Expected to find {index}"),
+         call. = FALSE)
+
+  header <- rmarkdown::yaml_front_matter(index)
+
+  if (is.null(header$site)) return(FALSE)
+
+  if (header$site == "workflowr::wflow_site") return(TRUE)
+
+  return(FALSE)
+}